fix(boinc): remove extraneous kaktoos validator
Pull #36 was meant to go in long before Pull #43.
Unfortunately I did not foresee both the length at which we'd be waiting on these to go in.
By the time we were gearing up to put these in, I probably should have cancelled pull #43, and merged the collection of those 2 PRs into a single PR both both commit histories.
The issue is that Pull #43 did not overwrite the kaktoos validator text set in Pull #36.
This should correct that.